On error while creating tasks, report the offending task.
This is especially applicable if tasks are defined using the create_doit_tasks() function
mechanism, which can have more complicated instructions and cause errors. If the
exception further occurs in a function that cannot be assigned to a task (e.g. a helper
function in use by multiple tasks), having the task name in the error message is helpful.